+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.7.4** 
+* Add `NamedProvider` for `CacheService` #542
 * `ProjectLayout` - provide layout for gradle_groovy combination #541
 * Cannot run a groovy program even in PROD mode #540
 * Implement new response output model #539